+++ app/src/features/apiClient/screens/apiClient/components/clientView/components/response/EmptyResponsePlaceholder/emptyResponsePlaceholder.scss
